requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
asteval==1.0.8
asttokens==2.4.1
astunparse==1.6.3
attrs==26.1.0
blinker==1.9.0
boto3==1.42.73
botocore==1.42.73
bw-migrations==0.2
bw2calc==2.4.0
bw2data==4.6
bw2io==0.9.14
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
certifi==2026.2.25
channels==4.3.1
channels_redis==4.3.0
chardet==7.2.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
comm==0.2.2
constructive_geometries==1.0
country_converter==1.3.2
crispy-bootstrap5==2026.3
dataflows-tabulator==1.54.3
datapackage==1.15.4
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.3.1
Django==5.2.7
django-crispy-forms==2.6
django-environ==0.13.0
django-extensions==4.1
django-phonenumber-field==8.3.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.0.1
flexcache==0.3
flexparser==0.4
fsspec==2026.2.0
greenlet==3.3.2
idna==3.11
ijson==3.5.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.7.2
jedi==0.19.1
jmespath==1.1.0
jsonlines==4.0.0
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.6.2
jupyter_core==5.7.2
linear-tsv==1.1.0
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
matplotlib-inline==0.1.7
matrix_utils==0.7.1
morefs==0.2.2
mrio-common-metadata==0.2.1
msgpack==1.1.2
nest-asyncio==1.6.0
numba==0.64.0
numpy==1.26.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==26.0
pandas==3.0.1
parso==0.8.4
peewee==4.0.2
phonenumbers==9.0.17
pillow==12.0.0
Pint==0.25.3
platformdirs==4.2.2
prettytable==3.17.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyecospold==4.0.1
PyJWT==2.12.1
pyparsing==3.3.2
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pyxlsb==1.0.10
PyYAML==6.0.3
pyzmq==26.1.0
qrcode==8.2
randonneur==0.7.2
randonneur_data==0.7.2
RapidFuzz==3.14.3
rdflib==7.6.0
redis==7.0.0
referencing==0.37.0
requests==2.32.5
rfc3986==2.0.0
rpds-py==0.30.0
s3transfer==0.16.0
scipy==1.17.1
setuptools==80.9.0
six==1.17.0
snowflake-id==1.0.2
SPARQLWrapper==2.0.0
sparse==0.18.0
SQLAlchemy==2.0.48
sqlparse==0.5.3
stack-data==0.6.3
stats_arrays==1.0
stripe==15.0.0
structlog==25.5.0
tableschema==1.21.0
toolz==1.1.0
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unfold==1.2.4
unicodecsv==0.14.1
urllib3==2.6.3
voluptuous==0.16.0
wcwidth==0.2.13
wheel==0.45.1
win32_setctime==1.2.0
wrapt==2.1.2
wurst==0.5.3
xarray==2026.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
